Added support for a * wildcard to clear up confusion
As discussed here:
https://github.com/MicrosoftDocs/windows-itpro-docs/issues/4408, the `.`
and `..` syntax are misleading and confusing at best. What they seem to
mean in practice is:
  - `.google.com` = Allow any site ending in "google.com"
  - `..google.com` = Allow only sub domains of "google.com"

This easily leads to awkward situations where:
  - "notgoogle.com" is allowed
  - "mail.google.com" is allowed, but not "google.com"

To mitigate this confusion (which leads to dangerous misconfigurations),
this commit adds support for the `*.` syntax. When this is used (i.e.
`*.google.com`), both "google.com" and "..google.com" are added to the
whitelists. The result is that "google.com" and "www.google.com" are
allowed, but not "notgoogle.com".